Overall architecture changes.   - Sandbox no longer needs a main function, just to create an inherited Application Instance and implement CreateApplication() function   - Created Layer stack so that sandbox can add layers to the behind application for things that will have no interaction with eachother   - Added new cmake section for automatically updating source_group visual studio structure   - Added precompiled header using CMake target_precompiled_headers function

+++ Engine/src/gfx/Shader.cpp
@@ -98,8 +98,8 @@ void Shader::ReadShaderFromFile(const std::string& path, ShaderSrc& src)
     }
 
     // TODO: Move this into the line search 
-    int vertStart = fileSrc.rfind("#vertex");
-    int fragStart = fileSrc.rfind("#fragment");
+    size_t vertStart = fileSrc.rfind("#vertex");
+    size_t fragStart = fileSrc.rfind("#fragment");
     
     if(vertStart < fragStart)
     {
